mindmemo/backend/pyproject.toml
[project]
name = "mindmemo-backend"
version = "1.0.0"
description = "MindMemo mental health companion backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
